requirements.txt
autogen
python-dotenv
praw